third_party/blink/web_tests/platform/generic/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-fallback-format-expected.txt

Checks incomplete datetime formats are rejected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Valid format
PASS getUserAgentShadowTextContent(input) is "1999-07-31 11:59 PM"
Invalid format, fallback to yyyy-MM-dd'T'HH:mm
PASS getUserAgentShadowTextContent(input) is "1999-07-31T23:59"
PASS getUserAgentShadowTextContent(input) is "1999-07-31T23:59"
PASS getUserAgentShadowTextContent(input) is "1999-07-31T23:59"
PASS successfullyParsed is true

TEST COMPLETE
